+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixes and improvements
 * chained translation tasks are now dispatched on multiple gpus
+* (worker) retry 5 times failed status, and correctly terminate
 
 ## [v1.0.0](https://github.com/OpenNMT/nmt-wizard/releases/tag/v1.0.0) (2018-10-04)